+++ RaiderEngine/demos/demoBrickBreaker/Paddle.cpp
@@ -12,13 +12,13 @@ Paddle::Paddle(glm::vec2 position) : GameObject2D(position, 0, glm::vec2(1), glm
 }
 
 void Paddle::restart() {
-	setCenter(SCR_WIDTH / 2, SCR_HEIGHT - 80);
+	setCenter(SCR_WIDTH / 2.f, SCR_HEIGHT - 80.f);
 }
 
 void Paddle::update() {
 	if (paused)
 		return;
 	translate((keyHeld("mvRight") - keyHeld("mvLeft")) * speed * deltaTime * levelMultiplier,0);
-	setCenter(std::fmax(0, std::fmin(SCR_WIDTH, center.x)), center.y);
+	setCenter(std::fmaxf(0, std::fminf(static_cast<float>(SCR_WIDTH), center.x)), center.y);
 }
 #endif
